backend/requirements.txt
asgiref==3.7.2
boto3==1.34.35
botocore==1.34.35
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.4
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
